refactor: centralize models and utils
