+++ app/routes/admin_debug.py
@@ -1,12 +1,12 @@
 from fastapi import APIRouter, Request, Depends, HTTPException, Form
 from fastapi.responses import RedirectResponse
-from app.utils.templates import templates
+from core.utils.templates import templates
 from sqlalchemy.orm import Session
 from typing import Optional
 
-from app.utils.db_session import get_db
-from app.utils.auth import require_role
-from app.models.models import AuditLog, User, Device
+from core.utils.db_session import get_db
+from core.utils.auth import require_role
+from core.models.models import AuditLog, User, Device
 from app.tasks import (
     start_trap_listener,
     stop_trap_listener,
