+++ app/routes/admin_users.py
@@ -3,11 +3,11 @@ from fastapi.responses import RedirectResponse
 from sqlalchemy.orm import Session
 import secrets
 
-from app.utils.db_session import get_db
-from app.utils.auth import require_role, get_password_hash, ROLE_CHOICES
-from app.utils.templates import templates
-from app.utils.audit import log_audit
-from app.models.models import User
+from core.utils.db_session import get_db
+from core.utils.auth import require_role, get_password_hash, ROLE_CHOICES
+from core.utils.templates import templates
+from core.utils.audit import log_audit
+from core.models.models import User
 
 router = APIRouter()
