+++ app/routes/audit.py
@@ -1,10 +1,10 @@
 from fastapi import APIRouter, Request, Depends
-from app.utils.templates import templates
+from core.utils.templates import templates
 from sqlalchemy.orm import Session
 
-from app.utils.db_session import get_db
-from app.utils.auth import require_role
-from app.models.models import AuditLog
+from core.utils.db_session import get_db
+from core.utils.auth import require_role
+from core.models.models import AuditLog
