+++ app/routes/inventory.py
@@ -1,9 +1,9 @@
 from fastapi import APIRouter, Request, Depends
 from sqlalchemy.orm import Session
-from app.utils.templates import templates
-from app.utils.auth import require_role
-from app.utils.db_session import get_db
-from app.models.models import Device, DeviceType
+from core.utils.templates import templates
+from core.utils.auth import require_role
+from core.utils.db_session import get_db
+from core.models.models import Device, DeviceType
 
 router = APIRouter()
