+++ app/routes/locations.py
@@ -2,10 +2,10 @@ from fastapi import APIRouter, Request, Depends, HTTPException, Form
 from fastapi.responses import RedirectResponse
 from sqlalchemy.orm import Session
 
-from app.utils.db_session import get_db
-from app.utils.auth import require_role
-from app.utils.templates import templates
-from app.models.models import Location
+from core.utils.db_session import get_db
+from core.utils.auth import require_role
+from core.utils.templates import templates
+from core.models.models import Location
 
 LOCATION_TYPES = ["Fixed", "Remote", "Mobile"]
