+++ app/routes/network.py
@@ -1,10 +1,10 @@
 from fastapi import APIRouter, Request, Depends, HTTPException
-from app.utils.templates import templates
+from core.utils.templates import templates
 from sqlalchemy.orm import Session
 
-from app.utils.db_session import get_db
-from app.utils.auth import get_current_user
-from app.models.models import Device
+from core.utils.db_session import get_db
+from core.utils.auth import get_current_user
+from core.models.models import Device
