+++ app/routes/port_config_templates.py
@@ -3,10 +3,10 @@ from datetime import datetime
 from fastapi.responses import RedirectResponse
 from sqlalchemy.orm import Session
 
-from app.utils.db_session import get_db
-from app.utils.auth import require_role
-from app.utils.templates import templates
-from app.models.models import PortConfigTemplate
+from core.utils.db_session import get_db
+from core.utils.auth import require_role
+from core.utils.templates import templates
+from core.models.models import PortConfigTemplate
 
 router = APIRouter()
