+++ app/routes/snmp_traps.py
@@ -1,11 +1,11 @@
 from datetime import datetime
 from typing import Optional
 from fastapi import APIRouter, Request, Depends
-from app.utils.db_session import get_db
+from core.utils.db_session import get_db
 from sqlalchemy.orm import Session
-from app.utils.auth import require_role
-from app.utils.templates import templates
-from app.models.models import SNMPTrapLog, Device
+from core.utils.auth import require_role
+from core.utils.templates import templates
+from core.models.models import SNMPTrapLog, Device
 
 router = APIRouter()
