+++ app/routes/ssh_tasks.py
@@ -6,12 +6,12 @@ import requests
 import asyncio
 from datetime import datetime
 
-from app.utils.db_session import get_db
-from app.utils.auth import require_role, user_in_site
-from app.models.models import Device, ConfigBackup, SystemTunable
-from app.utils.ssh import build_conn_kwargs, resolve_ssh_credential
-from app.utils.device_detect import detect_ssh_platform
-from app.utils.templates import templates
+from core.utils.db_session import get_db
+from core.utils.auth import require_role, user_in_site
+from core.models.models import Device, ConfigBackup, SystemTunable
+from core.utils.ssh import build_conn_kwargs, resolve_ssh_credential
+from core.utils.device_detect import detect_ssh_platform
+from core.utils.templates import templates
 
 router = APIRouter()
