+++ app/routes/syslog.py
@@ -3,10 +3,10 @@ from typing import Optional
 from fastapi import APIRouter, Request, Depends
 from sqlalchemy.orm import Session
 
-from app.utils.db_session import get_db
-from app.utils.auth import require_role
-from app.utils.templates import templates
-from app.models.models import SyslogEntry, Device, Site
+from core.utils.db_session import get_db
+from core.utils.auth import require_role
+from core.utils.templates import templates
+from core.models.models import SyslogEntry, Device, Site
 
 router = APIRouter()
