+++ app/routes/tunables.py
@@ -1,12 +1,12 @@
 from fastapi import APIRouter, Request, Depends, Form, Query
-from app.utils.templates import templates
+from core.utils.templates import templates
 from fastapi.responses import RedirectResponse
 from sqlalchemy.orm import Session
 import subprocess
 
-from app.utils.db_session import get_db
-from app.utils.auth import require_role
-from app.models.models import SystemTunable
+from core.utils.db_session import get_db
+from core.utils.auth import require_role
+from core.models.models import SystemTunable
