+++ core/__init__.py
@@ -1,0 +1,27 @@
+"""Core utilities and data models for server modules."""
+
+from . import models
+from .schemas import ColumnSelection
+from .utils.auth import (
+    get_current_user,
+    require_role,
+    user_in_site,
+    get_user_site_ids,
+    get_password_hash,
+    verify_password,
+    ROLE_CHOICES,
+    ROLE_HIERARCHY,
+)
+
+__all__ = [
+    "models",
+    "ColumnSelection",
+    "get_current_user",
+    "require_role",
+    "user_in_site",
+    "get_user_site_ids",
+    "get_password_hash",
+    "verify_password",
+    "ROLE_CHOICES",
+    "ROLE_HIERARCHY",
+]
