+++ core/schemas.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class ColumnSelection(BaseModel):
+    """List of selected column names for table preferences."""
+
+    selected: list[str] = []
